refactor(frontend): decompose EntityDetailModal god-component
Split the 919-line EntityDetailModal into a lean ~208-line orchestrator
plus cohesive units, behaviour-preserving:

- types.ts        — shared interfaces (props, HostClassification, EntityStats, …)
- format.ts       — local formatting helpers (bytes/number/snap-time/badge hue)
- hooks/          — 6 data hooks: useEntityRole, useEntityStats, useEntityNote,
                    useEntityHistory, useHostClassification, useIpSnapshotHistory
- sections/       — 6 presentational components: RoleSection,
                    HostClassificationSection, EntityStatsSection,
                    SnapshotHistoryTable, CaptureHistoryTable, NotesTab

The shell now only wires hooks to sections and owns tab/nested-modal state.
Public API unchanged (index.ts barrel export is stable). Local formatBytes/
formatNumber kept verbatim (they differ subtly from utils/formatters) — to be
reconciled in the helper-dedup slice.

Verified: tsc --noEmit clean, docker compose build, Playwright exercise of
APPLICATION (stats/top-IPs/capture-history/notes/Esc) and IP
(role/device-classification/snapshot-history) modals — no console errors.
